app/requirements.txt
streamlit==1.45.0
python-dotenv==1.1.0
openai==1.78.1
azure-search-documents==11.6.0
azure-core==1.34.0
folium==0.19.5
streamlit-folium==0.25.0
requests==2.32.3
